requirements.txt
albumentations==1.3.0
matplotlib==3.6.3
numpy==1.23.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
pandas==1.5.0
pillow==9.2.0
pydicom==2.3.0
pyfeats==1.0.0
pytorch-lightning==1.9.1
scikit-learn==1.1.3
scipy==1.10.1
segmentation-models-pytorch==0.3.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6
timm==0.9.2
torchmetrics==0.11.0
torchsampler==0.1.2
tornado==6.2
tqdm==4.64.1